+++ src/BusinessLogic/CreatePolls/ICreatePollService.cs
@@ -5,8 +5,8 @@ namespace BusinessLogic.CreatePolls
     public interface ICreatePollService
     {
         Task CreateAsync(CreatePollRequest createPollRequest);
-        Task<CreatePollRequest> FindPendingAsync(long chatId, int userId);
+        Task<CreatePollRequest> FindPendingAsync(long chatId, long userId);
         Task SaveAsync(CreatePollRequest pendingRequest);
-        Task<CreatePollRequest> FindPendingAndValidAsync(int fromId);
+        Task<CreatePollRequest> FindPendingAndValidAsync(long fromId);
     }
 }
